Moved GORM utils to utils_gorm.go
